글 수정 & 삭제 권한 체크

+++ src/main/java/com/rest1/domain/post/post/controller/ApiV1PostController.java
@@ -52,9 +52,17 @@ public class ApiV1PostController {
     @Operation(summary = "글 삭제")
     @DeleteMapping("/{id}")
     public RsData<Void> deleteItem(
-            @PathVariable Long id
+            @PathVariable Long id,
+            @RequestHeader("Authorization") @NotBlank @Size(min=30, max=50) String apiKey
     ) {
+
+        String authorization = apiKey.replace("Bearer ", "");
+        Member actor = memberService.findByApiKey(authorization).orElseThrow(() -> new ServiceException("401-1", "API 키가 올바르지 않습니다."));
+
         Post post = postService.findById(id).get();
+
+        //권한 체크
+        if(!actor.equals(post.getAuthor())) throw new ServiceException("403-1", "삭제 권한이 없습니다.");
         postService.delete(post);
 
         return new RsData<Void>(
@@ -81,24 +89,18 @@ public class ApiV1PostController {
     }
 
     @PostMapping
-    @Operation(summary = "글 작성")
     @Transactional
+    @Operation(summary = "글 작성")
     public RsData<PostWriteResBody> createItem(
             @RequestBody @Valid PostWriteReqBody reqBody,
-            @RequestHeader("Authorization") @NotBlank @Size(min=30, max=50)  String apiKey
+            @RequestHeader("Authorization") @NotBlank @Size(min=30, max=50) String apiKey
     ) {
+
+        //인증
         String authorization = apiKey.replace("Bearer ", "");
 
-        Member actor = memberService.findByApiKey(apiKey)
-                .orElseThrow(()
-                        -> new ServiceException("404-1", "API 키가 올바르지 않습니다."));
+        Member actor = memberService.findByApiKey(authorization).orElseThrow(() -> new ServiceException("401-1", "API 키가 올바르지 않습니다."));
         Post post = postService.write(actor, reqBody.title, reqBody.content);
-
-        //회원 사칭하는 것을 막기 위해 비밀번호 검증
-//        if(!actor.getPassword().equals(password))
-//            throw new ServiceException(
-//                    "401-1",
-//                    "비밀번호가 일치하지 않습니다.");
 
         return new RsData<>(
                 "201-1",
@@ -126,10 +128,19 @@ public class ApiV1PostController {
     @Transactional
     public RsData<Void> modifyItem(
             @PathVariable Long id,
-            @RequestBody @Valid PostModifyReqBody reqBody
+            @RequestBody @Valid PostModifyReqBody reqBody,
+            @RequestHeader("Authorization") @NotBlank @Size(min=30, max=50) String apiKey
+
     ) {
 
+        String authorization = apiKey.replace("Bearer ", "");
+        Member actor = memberService.findByApiKey(authorization).orElseThrow(() -> new ServiceException("401-1", "API 키가 올바르지 않습니다."));
         Post post = postService.findById(id).get();
+
+        //권한 체크
+        if(!actor.equals(post.getAuthor())) throw new ServiceException("403-1", "수정 권한이 없습니다.");
+
+        //수정 로직
         postService.modify(post, reqBody.title, reqBody.content);
 
         return new RsData(
